requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
gunicorn==21.2.0
psycopg2-binary>=2.9.10

# Discord Bot Dependencies (optional)
discord.py>=2.3.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
